Add lock to protect access to OpenBSC core
It will become necessary when MNCC messages will
have to be send from another thread.

Note that it's an easy but messy way. It would be
better to queue those cross thread MNCC messages
somewhere locally and process them in the main
monitor thread. But it works, it's easier, and can
be fairly easily changed afterwards ...

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>